Add files via upload
A java project developed using Swing Netbeans IDE.  This a project is a design tool that helps BPM designers in version and context definition, update, and storage. Modelled versions are recorded in the VP2M repository, which is an XML database (implemented with XDB), while contexts are recorded as query rules (i.e., SQWRL rules) in the domain ontology repository. Note that this ontology has been implemented with Protégé 5.1.0 ontology development environment, which is built on top of the OWL API, which is integrated with the Drools rule engine for supporting reasoning and inference, and which is compatible with both Java 7 and Java 8 versions. We have used OWL formalisations, OWL API and SWRL API to implement reasoning in the NetBeans IDE 8.1 environment.